Add files via upload
These files are a combination of different file formats. You can view my code respectively. The php is files are connecting the html (front-end) information and passing it on to a database for storage. In additional, the passed information gets displayed in a table on the front-end as well. This process is known as fetching.